+++ prisma/migrations/20260807145410_add_liq_pool_detailed_json_field/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "portfolio_snapshots" ADD COLUMN     "liq_pool_detailed" JSONB;